pdf_ex: Add assume-one-all
There are cases in the wild where an inner cell is not aligned properly,
especially for a cell that spans the entire column space (i.e they add
a new model and forget to re-align the stuff that's the same for all).

The --assume-one-all option ensures that this gets detected and the
value assigned (correctly) to all columns. I.e if there is only one
value for all columns (even if not centered correctly) assume it applies
to all columns.