tools: pipeline: Add support for multiple download/extract/map stages in one pipeline
This is especially useful for growatt, which have long range of product
in the same series and split it in two PDF files. It makes sense to use
a single pipeline, download a few datasheets, then extract all those and
map them.

At the moment the different stages don't share anything (i.e it's
a single "stage" per file, not a single stage operating on multiple
files), but for now it does the trick and it's more flexible.

So essentially what we can do now:

  download:
    datasheet_1: url1
    datasheet_2: url2
  extractor:
    - inherit: main
      source: filename_1  # this uses the first download and produces
      filename_1.csv
      area: [..]
    - inherit: main
      source: filename_2
      area: [..]
  mapper:
    - inherit: main
      source: filename_1  # this uses filename_1.csv and produces
      a number of outputs
    - inherit: main
      source: filename_2